Move visibility tracking to AutocompleteTelemetry class
